split gemstone out to it's own file.
