+++ src/include/coin.hpp
